requirements.txt
apispec==0.39.0
attrdict==2.0.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
dill==0.2.5
ecdsa==0.13
enum==0.4.7
Flask==0.12.2
flask-apispec==0.7.0
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonpickle==1.0
MarkupSafe==1.0
marshmallow==2.16.3
PyFunctional==1.2.0
PyYAML==3.13
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
tabulate==0.8.2
urllib3==1.22
virtualenv==16.0.0
webargs==4.1.2
Werkzeug==0.14.1
